Permit non-integral quadrature widths (oops)
git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@2780 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1

+++ src/omxExpectationBA81.cpp
@@ -517,7 +517,7 @@ static void ba81SetupQuadrature(omxExpectation* oo, int gridsize)
 	int numUnique = state->numUnique;
 	int numThreads = Global->numThreads;
 	int maxDims = state->maxDims;
-	int Qwidth = state->Qwidth;
+	double Qwidth = state->Qwidth;
 	int numSpecific = state->numSpecific;
 	int priDims = maxDims - (numSpecific? 1 : 0);
